refactor: Auth 예외 처리 중복 제거 및 전역 예외 핸들러 통합
